requirements.txt
# Core scraping and automation
playwright==1.40.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio==3.4.3

# LLM and AI
openai==1.6.1
anthropic==0.8.1
transformers==4.36.2
torch==2.1.2
huggingface-hub==0.19.4
onnxruntime==1.16.3

# Data processing and validation
pydantic==2.5.2
pandas==2.1.4
numpy==1.24.4

# Database and storage
pymongo==4.6.1
motor==3.3.2

# Google Sheets integration
gspread==5.12.4
oauth2client==4.1.3
google-auth==2.25.2

# Web utilities
requests==2.31.0
selenium==4.16.0
fake-useragent==1.4.0

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Proxy and anti-detection
python-socks==2.4.3
rotating-proxies==0.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.4

# Utilities
tqdm==4.66.1
rich==13.7.0
click==8.1.7
tenacity==8.2.3
email-validator==2.1.0
validators==0.22.0

# Additional dependencies for data processing
openpyxl==3.1.2
xlsxwriter==3.1.9